requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_gnu
aiofiles=0.7.0=pyhd3eb1b0_0
aiohttp=3.8.4=py311h459d7ec_1
aiosignal=1.3.1=pyhd8ed1ab_0
asttokens=2.2.1=pyhd8ed1ab_0
async-timeout=4.0.2=pyhd8ed1ab_0
attrs=23.1.0=pyh71513ae_1
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.5=pyhd8ed1ab_0
blas=1.0=openblas
bottleneck=1.3.5=py311hbed6279_0
brotlipy=0.7.0=py311h5eee18b_1002
bzip2=1.0.8=h7f98852_4
ca-certificates=2023.5.7=hbcca054_0
certifi=2023.5.7=pyhd8ed1ab_0
cffi=1.15.1=py311h5eee18b_3
charset-normalizer=3.2.0=pyhd8ed1ab_0
colorama=0.4.6=pyhd8ed1ab_0
cryptography=39.0.1=py311h9ce1e76_2
debugpy=1.5.1=py311h6a678d5_0
decorator=5.1.1=pyhd8ed1ab_0
executing=1.2.0=pyhd8ed1ab_0
freetype=2.12.1=hca18f0e_1
frozenlist=1.3.3=py311hd4cff14_0
giflib=5.2.1=h0b41bf4_3
idna=3.4=pyhd8ed1ab_0
importlib-metadata=6.8.0=pyha770c72_0
importlib_metadata=6.8.0=hd8ed1ab_0
ipykernel=6.15.0=pyh210e3f2_0
ipython=8.14.0=pyh41d4057_0
jbig=2.1=h7f98852_2003
jedi=0.18.2=pyhd8ed1ab_0
jpeg=9e=h0b41bf4_3
jupyter_client=8.3.0=pyhd8ed1ab_0
jupyter_core=5.3.1=py311h38be061_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.40=h41732ed_0
lerc=3.0=h295c915_0
libdeflate=1.8=h7f8727e_5
libexpat=2.5.0=hcb278e6_1
libffi=3.4.2=h7f98852_5
libgcc-ng=13.1.0=he5830b7_0
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=13.1.0=he5830b7_0
libnsl=2.0.0=h7f98852_0
libopenblas=0.3.21=h043d6bf_0
libpng=1.6.39=h753d276_0
libsodium=1.0.18=h36c2ea0_1
libsqlite=3.42.0=h2797004_0
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.3.0=h6f004c6_2
libuuid=2.38.1=h0b41bf4_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
libzlib=1.2.13=hd590300_5
lz4-c=1.9.3=h9c3ff4c_1
matplotlib-inline=0.1.6=pyhd8ed1ab_0
multidict=6.0.4=py311h2582759_0
ncurses=6.4=hcb278e6_0
nest-asyncio=1.5.6=pyhd8ed1ab_0
numexpr=2.8.4=py311hde40847_1
numpy=1.25.0=py311h24aa872_0
numpy-base=1.25.0=py311hbfb1bba_0
openssl=3.1.1=hd590300_1
packaging=23.1=pyhd8ed1ab_0
pandas=1.5.3=py311hba01205_0
parso=0.8.3=pyhd8ed1ab_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=9.4.0=py311h6a678d5_0
pip=23.1.2=pyhd8ed1ab_0
platformdirs=3.8.1=pyhd8ed1ab_0
prompt-toolkit=3.0.39=pyha770c72_0
prompt_toolkit=3.0.39=hd8ed1ab_0
psutil=5.9.5=py311h2582759_0
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.15.1=pyhd8ed1ab_0
pyopenssl=23.0.0=py311h06a4308_0
pysocks=1.7.1=py311h06a4308_0
python=3.11.4=hab00c5b_0_cpython
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.11=3_cp311
pytz=2021.3=pyhd3eb1b0_0
pyzmq=25.1.0=py311h6a678d5_0
readline=8.2=h8228510_1
requests=2.29.0=py311h06a4308_0
setuptools=68.0.0=pyhd8ed1ab_0
six=1.16.0=pyhd3eb1b0_1
stack_data=0.6.2=pyhd8ed1ab_0
tenacity=8.2.2=pyhd8ed1ab_0
tk=8.6.12=h27826a3_0
tornado=6.3.2=py311h459d7ec_0
tqdm=4.65.0=pyhd8ed1ab_1
traitlets=5.9.0=pyhd8ed1ab_0
typing-extensions=4.7.1=hd8ed1ab_0
typing_extensions=4.7.1=pyha770c72_0
tzdata=2023c=h71feb2d_0
urllib3=1.26.8=pyhd3eb1b0_0
wcwidth=0.2.6=pyhd8ed1ab_0
wheel=0.40.0=pyhd8ed1ab_0
xz=5.2.6=h166bdaf_0
yarl=1.9.2=py311h459d7ec_0
zeromq=4.3.4=h9c3ff4c_1
zipp=3.16.0=pyhd8ed1ab_1
zlib=1.2.13=hd590300_5
zstd=1.5.2=h8a70e8d_1
